cleanup image alignment scripts
